requirements.txt
PyQt6
mysql-connector-python
opencv-python
imutils
tensorflow
scipy
scikit-learn
imageio
pandas